requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
pycocotools>=2.0.6

# Scientific computing
scipy>=1.10.0
numpy>=1.24.0

# Image processing
Pillow>=9.5.0
opencv-python>=4.7.0
scikit-image>=0.20.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
tensorboard>=2.13.0
matplotlib>=3.7.0
optuna>=3.6.0
